+++ plugins/flux/src/components/page/PageList.tsx
@@ -57,7 +57,7 @@ const PageList: React.FunctionComponent<IPageListProps> = ({
     },
   );
 
-  // refetchhWithDelay is used to call the refetch function to get the resource, but with a delay of 3 seconde. This is
+  // refetchhWithDelay is used to call the refetch function to get the resource, but with a delay of 3 seconds. This is
   // required, because sometime the Kubenretes isn't that fast after an action (edit, delete, ...) was triggered.
   const refetchhWithDelay = (): void => {
     setTimeout(() => {